+++ CHANGELOG.md
@@ -8,12 +8,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 <!--
 
-## Unreleased: v14.5.0 (2026-07-28)
+## Unreleased: v14.5.0 (2026-07-29)
 
 ### Features
 
-- **Core**:
-- **Docs**: 
+- **Core**: 
+- **Docs**: Update Assistant Knowledge Base articles
+- **Docs**: Add authority level count table to dispatch generator
 
 ### Fixes